+++ backend/models/__init__.py
@@ -5,19 +5,20 @@ and their relationships. It also includes common functionality like
 timestamp tracking and soft delete support.
 """
 
-# First import base classes
-from backend.models.base import Base, SoftDeleteMixin, TimestampMixin
-from backend.models.booking import Booking, BookingStatus, PaymentStatus
+# Then import models with dependencies
+from backend.models.barcode_sequence import BarcodeSequence
 
 # Then import models without dependencies or with minimal dependencies
+from backend.models.booking import Booking, BookingStatus, PaymentStatus
 from backend.models.category import Category
 from backend.models.client import Client, ClientStatus
+
+# First import from core and mixins packages
+from backend.models.core import Base
 from backend.models.document import Document, DocumentStatus, DocumentType
 from backend.models.equipment import Equipment, EquipmentStatus
-
-# Then import models with dependencies
+from backend.models.mixins import SoftDeleteMixin, TimestampMixin
 from backend.models.subcategory_prefix import SubcategoryPrefix
-from backend.models.barcode_sequence import BarcodeSequence
 from backend.models.user import User
 
 __all__ = [
